:books: add the flask_jwt_extended for the login and logout functionality and also other decoratoers

+++ project/resources/item.py
@@ -1,4 +1,10 @@
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+    jwt_required,
+    get_jwt_claims,
+    jwt_optional,
+    get_jwt_identity,
+    fresh_jwt_required
+)
 from flask_restful import reqparse, Resource
 from project.models.item import ItemModel
 
@@ -16,7 +22,7 @@ class Item(Resource):
                         help="Every item needs a store id"
                         )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -24,6 +30,7 @@ class Item(Resource):
 
         return {"message": "Item not found"}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {"message": f"An item with name {name} already exists"}, 400
@@ -36,7 +43,11 @@ class Item(Resource):
             return {"message": "An error occurred inserting the item"}, 500
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims["is_admin"]:
+            return {"message": "Admin privilege required!"}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -60,6 +71,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        items = ItemModel.get_items()
-        return {'items': list(map(lambda x: x.json(), items))}, 200
+        user_id = get_jwt_identity()
+        items = [x.json() for x in ItemModel.get_items()]
+        if user_id:
+            return {'items': items}, 200
+        return {
+            "items": [item['name'] for item in items],
+            "message": "more data available if you log in"
+            }, 200
